+++ gateway/config/production.lua
@@ -5,4 +5,5 @@ return {
     configuration_cache = os.getenv('APICAST_CONFIGURATION_CACHE') or 5*60,
     timer_resolution = '100ms',
     port = { metrics = 9421 },
+    log_buffer_size = os.getenv("APICAST_ACCESS_LOG_BUFFER") or nil
 }